requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
SQLAlchemy==2.0.46
pymysql==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
APScheduler==3.10.4
